Exclude files when building the site
This commit adds a list of files to the configuration that should be exluded from site builds. This is creating a problem on Federalist where the node modules are being built into the site.

This was noticed when sites weren't updating because they were timing out after taking a long time to upload.