Update `link` shortcode
There's a lot packaged in here. First, I deprecated the use of `relref` as a shortcode, in favor of the homegrown `link` to allow a single shortcode to be used for multiple link types. Second, I have a new technique for dev mode that uses a dev config file accessed with `hugo serve --config=config_dev.yml` — this also sets a custom .Site.Param that the `link` shortcode uses to ignore links to ignored files.