Update 2021-02-08-call-for-participation-2021-government-ux-summit.md
Updating note section: The February 26th deadline for submissions has passed. Registration for the 2021 UX Summit will be open soon. We have received a number of submissions on a wide variety of topics related to UX! We will send a notification to the User Experience community once the registration is open. The summit will be held on June 22-24, 2021.