Update plainlanguage-may-event-title-card.yml
Added alt text: Title card image of multi-colored diagonal shapes with the lines of text on a black background. The first line has Positioning Yourself as a Plain Language Editor in white text. The second line has May 19 2021 in yellow text. The Digitalgov logo in white is below the text.