Update plainlanguage-july-event-title-card.yml
Included alt text: "Title card image of multi-colored diagonal shapes with the lines of text on a black background. The first line has Using Plain Language To Tell Our Stories in white text. The second line has July 14, 2021 in yellow text. The Digitalgov logo in white is below the text."