Update dap-july-event-title-card.yml
Adding the alt text to the title card: Title card image of multi-colored diagonal shapes with the lines of text on a black background. The first line has FedRAMP & DAP Case Study Part 1: Assessing Analytics Strategy in white text. The second line has July 28, 2021 in yellow text. The Digitalgov logo in white is below the text.