Added "Acronym"
Because the U.S. Copyright Office doesn't have an acronym, I didn't originally include an acronym in the workflow. But it looks inconsistent with the other authors on the SocialGov Summer event page, so I added "Copyright Office" to the acronym line so least something appears on the event page.