Update plain-language-september-event-title-card.yml
Adding the following alt text to the title card: Title card image of multi-colored diagonal shapes with the lines of text on a black background. The first line has Fourth Annual Federal Plain Language Summit in white text. The second line has September 21 2021 in yellow text. The Digitalgov logo in white is below the text.