Update 2022-plain-language-summit-title-card.yml
Adding alt text: Title card graphic for a summit has a black background with multi-colored diagonal shapes on the right side, and on the left, three lines of text and the Digital.gov logo. In yellow text, the first line says, Plain Language Community Of Practice. The second line has the title of the event, Federal Plain Language Summit 2022, in white bold text. The third line has Join the best and brightest plain language practitioners from across the government, in yellow text. In the bottom left corner is the white Digital.gov logo which consists of three ribbons joined together in an hexagon to resemble the shape of a capital letter D, followed by the domain and program name, Digital.gov.