Update anatomy-of-an-email-slide-federal-social-media-week.yml
remove caption, shorten alt text

original: "A slide from the December 2020 Federal Social Media Week presentation. On the left is an infographic that is titled, The Anatomy of a Good Email. It shows the outline of a human figure with three main sections, Header, Body, and Footer. In each, there are three to five subsections that have icons and short text descriptions. Click the link in the caption below this image for the full transcript. On the right of the slide is a screen capture of a GSA IT Insider email with eight email elements highlighted by red buttons; email header, fun subject line, use of an emoji in the subject line, branding with the newsletter's logo, an animated gif accompanies the message text, a call-to-action button, contact info, and a section to provide feedback."